broadcast and health checks
